requirements.txt
torch>=2.0.0
torchvision>=0.15.0
facenet-pytorch>=2.5.3
mtcnn>=0.1.1
scikit-learn>=1.3.0
matplotlib>=3.7.0
tqdm>=4.65.0
Pillow>=10.0.0
numpy>=1.24.0